Fix error where a bad folder would cause bad db entries
An exception when listing the subfolders of a bad folder (e.g. a broken
link) would cause the next folder to not be returned from
TreeTraverser#NextFolder. That meant that it wasn't registered as a new
parent folder, and its children didn't reference it in the DB.

In turn, this caused the folder to be purged from the filesystem after the
backup was complete.