Properly support multiple backends
Previously, a changed file would get updated in one backend, but not the
other. This would be due to the first backend updating the database,
so the second run thinking that there was nothing to do.

We now use the backends side-by-side, which is perhaps less ideal, but still
works fine.