In one fell swoop, eliminate pruning and testing
Instead, for unmodified files, we check whether the file exists
on the backend, and optionally whether it's been modified. This
also eliminates problems with pruning, where one backend would
report the file missing, and the other would therefore have its
file re-added.

Also, testing for MirrorBackend is always on because it's so cheap.
There's an option for testing S3 backups, which is stored in settings.
Because of this, there's no longer a 'Test & Backup' button