Rewrite restore operation so uses backend filesystem instead
This allows the files/folders tables to be missing (e.g. S3Backend) and the files will
still be restord. The database will also be populate.

The upshot is that files on the dest filesystem (where we're restoring to) will need
to be touched to match the mtime stored in the DB / on the backend